Move module name constant to index
