+++ web/client-js/src/store/board/types.ts
@@ -25,11 +25,6 @@
  */
 
 /**
- * Module name.
- */
-export const BOARD = "BOARD";
-
-/**
  * Aliases for frequently used types.
  */
